datatype double to int

+++ Scripts/Enemies/EnemyFlying.cs
@@ -6,7 +6,7 @@ namespace Sands
 {
     public class EnemyFlying : Enemy
     {
-        public EnemyFlying(int damage, double critChance, int maxHealth, int currentHealth) : base(damage, critChance, maxHealth, currentHealth) { }
+        public EnemyFlying(int damage, int critChance, int maxHealth, int currentHealth) : base(damage, critChance, maxHealth, currentHealth) { }
 
         //What is special about mooks?
     }
